+++ src/Autocomplete/src/EntityAutocompleterInterface.php
@@ -19,11 +19,6 @@ use Symfony\Bundle\SecurityBundle\Security;
  * Interface for classes that will have an "autocomplete" endpoint exposed.
  *
  * @template T of object
- *
- * TODO Remove next lines for Symfony UX 3
- *
- * @method array getAttributes(object $entity) Returns extra attributes to add to the autocomplete result.
- * @method mixed getGroupBy()                  Return group_by option.
  */
 interface EntityAutocompleterInterface
 {
@@ -58,9 +53,11 @@ interface EntityAutocompleterInterface
     /**
      * Returns extra attributes to add to the autocomplete result.
      *
-     * TODO Uncomment for Symfony UX 3
+     * @param T $entity
+     *
+     * @return array<string, mixed>
      */
-    /* public function getAttributes(object $entity): array; */
+    public function getAttributes(object $entity): array;
 
     /**
      * Return true if access should be granted to the autocomplete results for the current user.
@@ -69,10 +66,10 @@ interface EntityAutocompleterInterface
      */
     public function isGranted(Security $security): bool;
 
-    /*
+    /**
      * Return group_by option.
      *
-     * TODO Uncomment for Symfony UX 3
+     * @return string|null
      */
-    /* public function getGroupBy(): mixed; */
+    public function getGroupBy(): mixed;
 }